tweaks (#257)
* rename directory to playwright

* move pages/ one directory upwards

* move fixtures/ one directory upwards

* move app/ one directory upwards

* move assertions/ one directory upwards

* move flows/ one directory upwards

* organize pages into folders

* rename files to kebab case

* add types/node as a dev dependency

* update imports in tests/

* update path in playwright.config.ts

* update gitignore

* configure html report

* update gitignore
