+++ .gitignore
@@ -3,3 +3,6 @@
 .dist
 node_modules
 package-lock.json
+test-results
+.features-gen/
+playwright-report
