+++ playwright/package.json
@@ -21,6 +21,7 @@
     "license": "ISC",
     "devDependencies": {
         "@playwright/test": "^1.60.0",
+        "@types/node": "^25.9.1",
         "playwright-bdd": "^8.5.1"
     },
     "directories": {
