+++ playwright/tsconfig.json
@@ -1,0 +1,7 @@
+{
+    "compilerOptions": {
+        "types": [
+            "node"
+        ]
+    }
+}
